Updates Button render logic
